requirements.txt
# LiveKit Agent Framework
livekit-agents>=1.1.7
livekit-plugins-openai>=1.1.5
livekit-plugins-deepgram>=1.0.0
livekit-plugins-silero>=1.0.0

# Core dependencies
python-dotenv>=1.0.1
pydantic>=2.10.4
pydantic-ai>=0.4.6
pyyaml>=6.0.1
aiohttp>=3.9.0

# API dependencies for diagnosis chat
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.12.0
huggingface-hub>=0.20.0

# Development dependencies
pytest>=8.3.4
pytest-asyncio>=0.25.0
ruff>=0.8.4